+++ src/HTTP/Responses/UnsupportedContentEncoding.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rubix\Server\HTTP\Responses;
+
+class UnsupportedContentEncoding extends Response
+{
+    /**
+     * @param string[] $acceptedContentEncodings
+     */
+    public function __construct(array $acceptedContentEncodings)
+    {
+        parent::__construct(415, [
+            'Accept-Encoding' => implode(', ', $acceptedContentEncodings),
+        ]);
+    }
+}