+++ ui/components/MainFooter.vue
@@ -1,7 +1,7 @@
 <template>
     <footer class="footer">
         <div class="content has-text-centered">
-            <p class="is-size-7">© {{ new Date().getFullYear() }} <a href="https://rubixml.com" target="_blank">Rubix ML</a></p>
+            <p class="is-size-7">© {{ new Date().getFullYear() }} Rubix ML</p>
         </div>
     </footer>
 </template>